Modal config added
I have implemented the essential configurations for the modal, which now include a close button and basic properties. Moreover, I've introduced a new function to both the navbar component and the home route. This function enables the page to scroll to the top automatically whenever there is a route change.