Promotion and Delivery sections added
I have made some enhancements to the website. Firstly, I have added a 'Deals and Promotions' section, which is now included in the order card. To make it easier for users to find pizzas that are on sale, I implemented a filtering system directly in the pizzaData.js file.

Additionally, I have introduced a delivery modal that provides users with the option to choose between 'Delivery' or 'Pickup' methods. This modal enhances the user experience by simplifying the selection process.